Refactored pkg/kubemodel to core/pkg/kubemodel
