+++ core/pkg/kubemodel/window.go
